src/main/java/ru/sevastopall/readersDairy/service/impl/SimpleAuthorService.java
package ru.sevastopall.readersDairy.service.impl;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import ru.sevastopall.readersDairy.model.Author;
import ru.sevastopall.readersDairy.repository.AuthorRepository;
import ru.sevastopall.readersDairy.service.AuthorService;

import java.time.LocalDate;
import java.util.List;

@RequiredArgsConstructor
@Service
public class SimpleAuthorService implements AuthorService {
    private final AuthorRepository authorRepository;

    @Override
    public Author save(Author author) {
        return authorRepository.save(author);
    }

    @Override
    public Author findById(long id) {
        return authorRepository.findById(id).get();
    }

    @Override
    public List<Author> findByFirstName(String firstName) {
        return authorRepository.findByFirstName(firstName);
    }

    @Override
    public Author findByLastName(String lastName) {
        return authorRepository.findByLastName(lastName);
    }

}
